mvoe scan components button to components tab
